Refactor my code
