+++ routes/forgotPasswordRoute.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const User = require('../models/User');
+const User = require('./User');
 const nodemailer = require('nodemailer');
 const crypto = require('crypto');
