+++ routes/resetPasswordRoute.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const User = require('../models/User');
+const User = require('./User');
 const bcrypt = require('bcrypt');
 
 router.get('/:token', async (req, res) => {
