+++ server.js
@@ -7,22 +7,27 @@ const nodemailer = require('nodemailer');
 const crypto = require('crypto');
 const ejs = require('ejs'); // need to import because it's 'require' (npm i ejs)
 // Routes
-const forgotPasswordRoute = require('./wellbot/routes/forgotPasswordRoute');
-const launchRoute = require('./wellbot/routes/launchRoute');
-const loginRoute = require('./wellbot/routes/loginRoute');
-const resetPasswordRoute = require('./wellbot/routes/resetPasswordRoute');
-const signupRoute = require('./wellbot/routes/signupRoute');
+const forgotPasswordRoute = require('./routes/forgotPasswordRoute');
+const launchRoute = require('./routes/launchRoute');
+const loginRoute = require('./routes/loginRoute');
+const resetPasswordRoute = require('./routes/resetPasswordRoute');
+const signupRoute = require('./routes/signupRoute');
 const app = express();
 const PORT = process.env.PORT || 3000;
-
 const blogRoute = require('./routes/blogRoute');
+
+//############################################Do not touch######################################################
+app.use('/css', express.static(path.join(__dirname, 'css'))); // Need this to access the css files. Do not remove.
+app.use('/images', express.static(path.join(__dirname, 'images'))); // Need this to access the css files. Do not remove.
+app.set('view engine', 'ejs');
+//####################################################################################################
 
 
 // Express Middleware
 app.use(express.urlencoded({ extended: true }));
 app.use(express.static('public'));
-app.set('view engine', 'ejs');
-app.set('views', path.join(__dirname, 'wellbot/html/templates'));
+// app.set('views', path.join(__dirname, 'wellbot/html/templates'));
+
 // MongoDB Connection
 mongoose.connect(process.env.MONGODB_URI, {
   useNewUrlParser: true,
@@ -59,17 +64,6 @@ app.use('/login', loginRoute);
 app.use('/reset-password', resetPasswordRoute);
 app.use('/signup', signupRoute);
 
-// Start Server
-app.listen(PORT, () => {
-  console.log(`Server listening on port ${PORT}`);
-});
-
-
-//############################################Do not touch######################################################
-app.use('/css', express.static(path.join(__dirname, 'css'))); // Need this to access the css files. Do not remove.
-app.use('/images', express.static(path.join(__dirname, 'images'))); // Need this to access the css files. Do not remove.
-app.set('view engine', 'ejs');
-//####################################################################################################
 
 //Richard's script for collapsing meals and exercises button on homepage
 app.get('/home', (req, res) => {
@@ -78,6 +72,8 @@ app.get('/home', (req, res) => {
 
 app.use(blogRoute);
 
-app.listen(3000, () => {
-    console.log(`Server is running on port 3000`);
+// Start Server
+app.listen(PORT, () => {
+  console.log(`Server listening on port ${PORT}`);
 });
+
