+++ views/launch.ejs
@@ -1,0 +1,31 @@
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+  <link rel="stylesheet" href="css/launch.css">
+</head>
+<body>
+  <div class="container">
+    <div class="logo-wrap">
+      <!-- <img src="images/Wellbot.png" alt="Wellbot Logo" class="main-logo" /> -->
+    </div>
+    <h1 class="brand-name">WELLBOT</h1>
+    <p class="tagline">AI-powered fitness, tailored to you.</p>
+  
+    <div class="button-container">
+      <form method="POST" action="/"> 
+        <button type="submit" name="action" value="getStarted" class="btn btn-primary">Get Started</button>
+      </form>
+      <form method="POST" action="/">
+        <button type="submit" name="action" value="login" class="btn btn-secondary">Log In</button>
+      </form>
+    </div>
+  </div>
+</body>
+</html>
+
+ 
+
