--- wellbot/html/templates/header.ejs
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>WellBot</title>
-    <link href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
-    <!-- <link href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-icons/1.8.1/font/bootstrap-icons.min.css" -->
-        rel="stylesheet">
-    <link href="/css/style.css" rel="stylesheet">
-</head>
-
-<body>
-    <header class="header text-white text-center py-3">
-        <div class="conatiner-fluid">
-            <h1>WellBot</h1>
-        </div>
-    </header>
